+++ src/routes/compare/Compare.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 
 import { useRoute } from 'wouter';
 
+import { retrieveComputePlans } from '@/modules/computePlans/ComputePlansSlice';
 import { loadSeries } from '@/modules/series/SeriesSlice';
 
 import useAppDispatch from '@/hooks/useAppDispatch';
@@ -11,6 +12,9 @@ import { useDocumentTitleEffect } from '@/hooks/useDocumentTitleEffect';
 import { PATHS } from '@/routes';
 
 import PerfBrowser from '@/components/PerfBrowser';
+import PerfSidebarSectionComputePlans from '@/components/PerfSidebarSectionComputePlans';
+import PerfSidebarSettingsAverage from '@/components/PerfSidebarSettingsAverage';
+import PerfSidebarSettingsNodes from '@/components/PerfSidebarSettingsNodes';
 
 const Compare = (): JSX.Element => {
     const dispatch = useAppDispatch();
@@ -24,12 +28,27 @@ const Compare = (): JSX.Element => {
     );
 
     useEffect(() => {
+        dispatch(retrieveComputePlans({ computePlanKeys: keys }));
         dispatch(loadSeries(keys));
     }, []);
 
     const loading = useAppSelector((state) => state.series.loading);
     const series = useAppSelector((state) => state.series.series);
+    const computePlans = useAppSelector(
+        (state) => state.computePlans.computePlans
+    );
 
-    return <PerfBrowser loading={loading} series={series} computePlan={null} />;
+    return (
+        <PerfBrowser
+            loading={loading}
+            series={series}
+            computePlans={computePlans}
+            settingsComponents={[
+                PerfSidebarSettingsNodes,
+                PerfSidebarSettingsAverage,
+            ]}
+            sectionComponents={[PerfSidebarSectionComputePlans]}
+        />
+    );
 };
 export default Compare;